fix name of ProfilePhotoWrapper.tsx to the UserActionsWrapper.tsx, and extracted ProfilePhotoWrapper from there
