+++ frontend/src/components/layout/ShoppingCart.tsx
@@ -1,0 +1,31 @@
+import { Badge, Link } from "@mui/material";
+import ShoppingCartIcon from "@mui/icons-material/ShoppingCart";
+import React from "react";
+
+const ShoppingCartMiniature = () => {
+  return (
+    <Link href="/cart" className="cart" sx={{ margin: "auto" }}>
+      <Badge
+        badgeContent={5} // Hardcoded value of items in the cart
+        color="secondary"
+        anchorOrigin={{
+          vertical: "bottom",
+          horizontal: "left",
+        }}
+        sx={{
+          "& .MuiBadge-badge": {
+            fontSize: "1rem",
+            height: "1.2rem",
+            minWidth: "1.2rem",
+            transform: "translate(-25%, 25%)",
+            backgroundColor: "#ea2323",
+          },
+        }}
+      >
+        <ShoppingCartIcon sx={{ fontSize: "2.5rem", color: "#159d17" }} />
+      </Badge>
+    </Link>
+  );
+};
+
+export default ShoppingCartMiniature;
